back to humannat
Apparently, \bibliographystyle{humannat} threw up errors in development of Issue 3, for unknown reasons. This was mentioned in the version history, but not present in the \bibliographystyle{}. command. The line has now been removed from the version history too, and our strategy is to test whether humannat still produces errors when we start preparing issue 4.

Additonally, a series of small, meaningless changes to improve (hopefully) the layout of the documentation using %comments.